Рефакторинг ХОК Снепшот тестов, выносит мок компонент в тестовые данные
