+++ src/hocs/with-active-player/with-active-player.test.js
@@ -1,24 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import PropTypes from "prop-types";
+import {MockComponent} from "../../test-data";
 import withActivePlayer from "./with-active-player";
-
-const MockComponent = (props) => {
-  const {children} = props;
-
-  return (
-    <div>
-      {children}
-    </div>
-  );
-};
-
-MockComponent.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node
-  ]).isRequired,
-};
 
 const MockComponentWrapped = withActivePlayer(MockComponent);
